% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{ERA5_day_hurs}
\alias{ERA5_day_hurs}
\title{ERA5 daily relative humidity}
\format{
A climate4R grid object
}
\source{
ERA5 reanalysis data
}
\description{
A climate4R grid with daily relative humidity (hurs) from ERA5.
}
\keyword{data}
