% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatindexGrid.R
\name{heatindexGrid}
\alias{heatindexGrid}
\title{Simplified Physiological Heat Index in climate4R}
\usage{
heatindexGrid(tas, hurs, parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{tas}{A climate4R dataset of air temperature.}

\item{hurs}{A climate4R dataset of relative humidity.}

\item{parallel}{Logical. Enable parallel processing (default = FALSE).}

\item{max.ncores}{Integer. Maximum number of cores to use (default = 16).}

\item{ncores}{Integer. Specific number of cores to use.}
}
\value{
A climate4R object with the computed index in Kelvin.
}
\description{
Computation of the simplified physiological heat index introduced by Lu et al. (2025) directly
from climate4R objects.
}
\details{
Air temperature is internally converted to Kelvin and relative humidity to 0-1 fraction.
The vectorised C++ backend \code{heatindex_vec} of the \pkg{heatindex} package (Lu and Romps, 2025)
is internally called.
}
\examples{
\dontrun{
library(climate4R.heatindex)

data("ERA5_day_t2m", package = "climate4R.heatindex")
data("ERA5_day_hurs", package = "climate4R.heatindex")

sphi <- heatindexGrid(tas = ERA5_day_t2m, hurs = ERA5_day_hurs)
}

}
\references{
Lu, Y.-C. and Romps, D. M. (2025). heatindex: Tools for Calculating Heat Stress. \url{https://heatindex.org}.

Lu, Y., A. Goodman, P. Kalmus, and D. M. Romps, 2026: Simpler and Faster: An Improved Heat Index.
J. Appl. Meteor. Climatol., 65, 665–679, https://doi.org/10.1175/JAMC-D-25-0067.1.
}
\author{
climate4R adaptation by C. Rodriguez-Rumayor.
Original \pkg{heatindex} by Yi-Chuan Lu & David M. Romps \url{https://heatindex.org}.
}
